TASK-29 1.tabs api expanded 2.minor style changes 3.front toolbar-viewing.component.tsx implemented
