+++ src/assets/styles/styles-all.scss
@@ -28,3 +28,4 @@
 @import "styles/custom/autocomplete";
 @import "styles/custom/management_page";
 @import "styles/custom/button";
+@import "styles/custom/route-buttons.module";
