+++ src/assets/styles/styles/custom/_toolbar.scss
@@ -2,28 +2,38 @@
   display: flex;
   flex-direction: row;
   width: 100%;
-  padding: 10px;
 
-  form {
-    margin-left: auto;
-    margin-right: 15px;
-    border: 1px solid rgba($black, 0.37);
-    border-radius: 5px;
-    padding: 15px;
-
-    * {
-      margin: 3px;
-    }
-
-    .submit-button {
-      margin-top: 10px;
-      font-size: 15px;
-    }
-
-    #submit-button-container {
-      display: flex;
-      flex-direction: row;
-      justify-content: flex-start;
-    }
+  .buttons {
+    display: flex;
+    flex-direction: row;
+    width: 100%;
+    margin: 0;
   }
+
+  .filler {
+    flex-grow: 1;
+  }
+
+  //form {
+  //  margin-left: auto;
+  //  margin-right: 15px;
+  //  border: 1px solid rgba($black, 0.37);
+  //  border-radius: 5px;
+  //  padding: 15px;
+  //
+  //  * {
+  //    margin: 3px;
+  //  }
+  //
+  //  .submit-button {
+  //    margin-top: 10px;
+  //    font-size: 15px;
+  //  }
+  //
+  //  #submit-button-container {
+  //    display: flex;
+  //    flex-direction: row;
+  //    justify-content: flex-start;
+  //  }
+  //}
 }
