+++ src/components/schedule-page/import-buttons/import-buttons.component.tsx
@@ -61,7 +61,7 @@ export function ImportButtonsComponent(): JSX.Element {
 
   return (
     <div>
-      <Button variant="primary" onClick={handleToggle} ref={anchorRef}>
+      <Button variant="outlined" onClick={handleToggle} ref={anchorRef}>
         Plik
         <ArrowDropDownIcon />
       </Button>
