+++ src/components/schedule-page/toolbar/toolbar-viewing.component.tsx
@@ -1,0 +1,43 @@
+import React from "react";
+import { Button } from "../../common-components";
+import { ImportButtonsComponent } from "../import-buttons/import-buttons.component";
+import RouteButtonsComponent from "../../common-components/route-buttons/route-buttons.component";
+import SchedulePage from "../schedule-page.component";
+import ManagementPage from "../../workers-page/management-page.component";
+import { Box } from "@material-ui/core";
+
+interface TabData {
+  label: string;
+  component: JSX.Element;
+  right_side_buttons: JSX.Element;
+}
+
+function Buttons(): JSX.Element {
+  return (
+    <>
+      <div className={"buttons"}>
+        <ImportButtonsComponent />
+        <Box>
+          <Button size="small" className="submit-button" variant="primary">
+            Edytuj
+          </Button>
+        </Box>
+      </div>
+    </>
+  );
+}
+
+const tabs: TabData[] = [
+  // eslint-disable-next-line @typescript-eslint/camelcase
+  { label: "Plan", component: <SchedulePage />, right_side_buttons: Buttons() },
+  // eslint-disable-next-line @typescript-eslint/camelcase
+  { label: "Zarządzanie", component: <ManagementPage />, right_side_buttons: <div /> },
+];
+
+export function ToolbarViewingComponent(): JSX.Element {
+  return (
+    <>
+      <RouteButtonsComponent tabs={tabs} />
+    </>
+  );
+}
